% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{beforeFilterPrj}
\alias{beforeFilterPrj}
\title{* `beforeFilterPrj` : number of reads that got filtered out}
\format{
List with "x", "var_pcs", "rotation"
}
\usage{
beforeFilterPrj
}
\description{
* `beforeFilterPrj` : number of reads that got filtered out
}
\keyword{datasets}
