% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanHistoryPath.R
\name{cleanHistoryPath}
\alias{cleanHistoryPath}
\title{clean SCHNAPPs history path of RData files that are not referenced in any Rmd file in that directory.}
\usage{
cleanHistoryPath(path, verbose = TRUE, dryrun = TRUE)
}
\arguments{
\item{path}{History path that holds min. one Rmd file}

\item{verbose}{Print what is/would be executed}

\item{dryrun}{don't execute the removal commands.}
}
\value{
nothing
}
\description{
clean SCHNAPPs history path of RData files that are not referenced in any Rmd file in that directory.
}
