% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{dbCluster}
\alias{dbCluster}
\title{* `dbCluster` : used clustering}
\format{
List with "x", "var_pcs", "rotation"
}
\usage{
dbCluster
}
\description{
* `dbCluster` : used clustering
}
\keyword{datasets}
