% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Main.R
\docType{data}
\name{.schnappsEnv}
\alias{.schnappsEnv}
\title{.schnappsEnv 
Environment for package}
\format{An object of class \code{environment} of length 0.}
\usage{
.schnappsEnv
}
\description{
.schnappsEnv 
Environment for package
}
\keyword{datasets}
