% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{geneCount}
\alias{geneCount}
\title{* `geneCount` : number of genes per cell.}
\format{
List with "x", "var_pcs", "rotation"
}
\usage{
geneCount
}
\description{
* `geneCount` : number of genes per cell.
}
\keyword{datasets}
