% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{pca}
\alias{pca}
\title{* `pca` : pca results, Eigenvalues and vectors.}
\format{
List with "x", "var_pcs", "rotation"
}
\usage{
pca
}
\description{
* `pca` : pca results, Eigenvalues and vectors.
}
\keyword{datasets}
