% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{scExLite}
\alias{scExLite}
\title{Example data for schnapps app}
\format{
A data frame with 53940 rows and 10 variables
}
\source{
<http://www.diamondse.info/>
}
\usage{
scExLite
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. The variables are as follows:
}
\details{
* `scExLite`: singlecellExperiment object
}
\keyword{datasets}
