% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\name{schnappsLite}
\alias{schnappsLite}
\title{shiny server relevant functions}
\usage{
schnappsLite(
  data = "RdataFile",
  localContributionDir = "~/Rstudio/shHubgit/Dummy/",
  defaultValueSingleGene = "CD52",
  defaultValueMultiGenes = "CD52, S100A4, S100A9, S100A8",
  defaultValueRegExGene = "",
  DEBUG = FALSE,
  DEBUGSAVE = FALSE,
  historyPath = NULL,
  maxCells = 3000
)
}
\arguments{
\item{data}{RdataFile data file with scEx, scEx_log, projections}

\item{localContributionDir}{path to the directory(ies) that contain additional functionality}

\item{defaultValueSingleGene}{single gene name to used as a default value.}

\item{defaultValueMultiGenes}{comma separated list of gene names to be used as a default value.}

\item{defaultValueRegExGene}{regular Expression used for gene selection}

\item{DEBUG}{TRUE/FALSE whether to show debugging information on the console}

\item{DEBUGSAVE}{TRUE/FALSE where or not save internal data (very time consuming)}

\item{historyPath}{location (directory) where history directories and data will be stored.
historyPath should be used to generate Report}

\item{maxCells}{maximal number of cells to use. Used to limit memory usage for server}
}
\description{
shiny server relevant functions
}
\details{
Shiny app for the analysis of single cell data
}
\examples{
packPath = "schnappsPackage"
# TODO fix:
# schnappsLite(data = paste0(packPath, "/data/scExLite.RData"))
}
