% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{scol}
\alias{scol}
\title{* `scol` : color definitions for samples}
\format{
a vector of colour values for samples
}
\usage{
scol
}
\description{
* `scol` : color definitions for samples
}
\keyword{datasets}
