% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{umapReact}
\alias{umapReact}
\title{* `umapReact` : umap projections}
\format{
List with "x", "var_pcs", "rotation"
}
\usage{
umapReact
}
\description{
* `umapReact` : umap projections
}
\keyword{datasets}
