% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Lite.R
\docType{data}
\name{ccol}
\alias{ccol}
\title{* `ccol` : color definitions}
\format{
a vector with colours used for clusters
}
\usage{
ccol
}
\description{
* `ccol` : color definitions
}
\keyword{datasets}
