% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlotwithModuleScore.R
\name{DotPlotwithModuleScore}
\alias{DotPlotwithModuleScore}
\title{DotPlotwithModuleScore}
\usage{
DotPlotwithModuleScore(
  object,
  assay = NULL,
  features,
  cols = c("lightgrey", "blue"),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  idents = NULL,
  group.by = NULL,
  split.by = NULL,
  cluster.idents = FALSE,
  scale = TRUE,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA
)
}
\arguments{
\item{features}{list of character vectors of gene names in "symbol" annotation (SCHNAPPs specific)}
}
\description{
adapted from Seurat::DotPlot but adds the module score per group
}
