% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Main.R
\name{schnapps}
\alias{schnapps}
\title{shiny server relevant functions}
\usage{
schnapps(localContributionDir = "~/Rstudio/shHubgit/Dummy/",
  defaultValueSingleGene = "CD52",
  defaultValueMultiGenes = "CD52, S100A4, S100A9, S100A8",
  defaultValueRegExGene = "", DEBUG = TRUE, DEBUGSAVE = FALSE)
}
\arguments{
\item{localContributionDir}{localContributionDir}

\item{defaultValueSingleGene}{defaultValueSingleGene}

\item{defaultValueMultiGenes}{defaultValueMultiGenes}

\item{defaultValueRegExGene}{defaultValueRegExGene}

\item{DEBUG}{DEBUG}

\item{DEBUGSAVE}{DEBUGSAVE}
}
\description{
shiny server relevant functions
}
\details{
Shiny app for the analysis of single cell data
}
