% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schnapps-Main.R
\name{schnapps}
\alias{schnapps}
\title{shiny server relevant functions}
\usage{
schnapps(
  localContributionDir = "~/Rstudio/shHubgit/Dummy/",
  defaultValueSingleGene = "CD52",
  defaultValueMultiGenes = "CD52, S100A4, S100A9, S100A8",
  defaultValueRegExGene = "",
  DEBUG = FALSE,
  DEBUGSAVE = FALSE,
  historyPath = NULL
)
}
\arguments{
\item{localContributionDir}{path to the directory(ies) that contain additional functionality}

\item{defaultValueSingleGene}{single gene name to used as a default value.}

\item{defaultValueMultiGenes}{comma separated list of gene names to be used as a default value.}

\item{defaultValueRegExGene}{regular Expression used for gene selection}

\item{DEBUG}{TRUE/FALSE whether to show debugging information on the console}

\item{DEBUGSAVE}{TRUE/FALSE where or not save internal data (very time consuming)}

\item{historyPath}{location (directory) where history directories and data will be stored.}
}
\description{
shiny server relevant functions
}
\details{
Shiny app for the analysis of single cell data
}
\examples{
# create example data
data("scEx", package = "SCHNAPPs")
save(file = "scEx.Rdata", list = "scEx")
# use "scEx.Rdata" with load data functionality within the shiny app
}
